[contour] More OpenGL related refactors, splitting more generic code out of the OpenGL related parts.
